Add files via upload
The following files are the updated JAR template files where the motors for upper intake are reversed back and the matchloading section of the auton is much smoother because of the decrease in drive-train voltage. I also added a motion in matchload_rigt auton() where the bot pushes the balls further back.